requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Configuration
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
email-validator>=2.0.0

# HTTP Client
requests==2.32.3

# Google Gemini
google-generativeai==0.8.3

# Utils
python-dateutil==2.8.2

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Whisper (local transcription)
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy<2.0.0,>=1.24.0  # PyTorch 2.1.0 requires NumPy < 2.0

# Deepgram (cloud transcription)
deepgram-sdk>=5.3.0
httpx>=0.27.0
websockets>=12.0

